[업데이트] 댓글 기능 및 좋아요 기능 구현

+++ contents/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, redirect
-from contents.models import Feed
+from contents.models import Feed, Comment
 from django.contrib.auth.decorators import login_required
 
 # Create your views here.
@@ -23,14 +23,20 @@ def FeedUload(request):
 @login_required
 def FeedDelete(request, id):
     feed = Feed.objects.get(id=id)
-    feed.delete()
-    return redirect(request.META['HTTP_REFERER'])
+    if feed.user == request.user:
+        feed.delete()
+        return redirect(request.META['HTTP_REFERER'])
+    else:
+        return redirect(request.META['HTTP_REFERER'])
 
 @login_required
 def FeedChange(request, id):
     if request.method == "GET":
         feed = Feed.objects.get(id=id)
-        return render(request, 'feed_Change.html', {"feed":feed})
+        if feed.user == request.user:
+            return render(request, 'feed_Change.html', {"feed":feed})
+        else:
+            return redirect(request.META['HTTP_REFERER'])
 
     if request.method == "POST":
         feed_change = Feed.objects.get(id=id)
@@ -41,3 +47,42 @@ def FeedChange(request, id):
         feed_change.save()
         
         return redirect('/')
+
+
+@login_required
+def likes(request, id):
+    if request.method == 'POST':
+        feed = Feed.objects.get(id=id)
+        if feed.like_authors.filter(id=request.user.id).exists():
+            feed.like_authors.remove(request.user)
+        else:
+            feed.like_authors.add(request.user)
+
+        return redirect(request.META['HTTP_REFERER'])
+        
+
+
+
+def write_comment(request, id): 
+    if request.method == 'POST':
+        current_comment = Feed.objects.get(id=id)
+        comment = request.POST.get('comment')
+
+        FC = Comment()
+        FC.comment = comment
+        FC.user = request.user
+        FC.feed = current_comment
+        FC.save()
+
+    return redirect(request.META['HTTP_REFERER'])
+
+
+def delete_comment(request, id): 
+        comment = Comment.objects.get(id=id)        
+        if comment.user == request.user:
+            comment.delete()
+            return redirect('/')
+        else:
+            return redirect('/')
+        
+
